# swagger.json
{
  "openapi": "3.0.1",
  "servers": [
    {
      "url": "https://github.com/samchon/nestia",
      "description": "insert your server url"
    }
  ],
  "info": {
    "version": "0.1.0",
    "title": "Generated by nestia - https://github.com/samchon/nestia"
  },
  "paths": {
...
    "/users/{loginID}": {
      "get": {
        "tags": [],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "description": "",
            "schema": {},
            "required": true
          },
          {
            "name": "loginID",
            "in": "path",
            "description": "",
            "schema": {
              "type": "string",
              "nullable": false
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userdata"
                }
              }
            }
          }
        },
        "description": ""
      }
    },
    "/users": {
      "get": {
        "tags": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "description": ""
      },
      "post": {
        "tags": [],
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalSignUpRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/Users"
                    },
                    {
                      "$ref": "#/components/schemas/BadRequestException"
                    }
                  ]
                }
              }
            }
          }
        },
        "description": ""
      }
    }
  },
  "components": {
    "schemas": {
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "nullable": false
          },
          "githubID": {
            "type": "number",
            "nullable": false
          },
          "email": {
            "type": "string",
            "nullable": false
          },
          "nickname": {
            "type": "string",
            "nullable": false
          },
          "loginID": {
            "type": "string",
            "nullable": false
          },
          "blog": {
            "type": "string",
            "nullable": false
          },
          "bio": {
            "type": "string",
            "nullable": false
          },
          "avatarUrl": {
            "type": "string",
            "nullable": false
          },
          "githubPageUrl": {
            "type": "string",
            "nullable": false
          },
          "positionType": {
            "type": "string",
            "enum": [
              "Front-End",
              "Back-End",
              "Full-Stack"
            ],
            "nullable": false
          },
          "loginType": {
            "type": "string",
            "enum": [
              "local",
              "github"
            ],
            "nullable": false
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ],
            "nullable": false
          },
          "createdAt": {
            "type": "string",
            "nullable": false
          },
          "updatedAt": {
            "type": "string",
            "nullable": false
          },
          "deletedAt": {
            "type": "string",
            "nullable": false
          }
        },
        "nullable": false,
        "required": [
          "id",
          "githubID",
          "email",
          "nickname",
          "loginID",
          "blog",
          "bio",
          "avatarUrl",
          "githubPageUrl",
          "positionType",
          "loginType",
          "role",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
...
      "BlogPosts": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "nullable": false
          },
          "title": {
            "type": "string",
            "nullable": false
          },
          "Tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlogPostsTags"
            },
            "nullable": false
          },
          "content": {
            "type": "string",
            "nullable": false
          },
          "thumbnail": {
            "type": "string",
            "nullable": false
          },
          "UserId": {
            "type": "number",
            "nullable": false
          },
          "User": {
            "$ref": "#/components/schemas/Users"
          },
          "LikeDisLike": {
            "$ref": "#/components/schemas/BlogPostsLike"
          },
          "createdAt": {
            "$ref": "#/components/schemas/Date"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Date"
          },
          "deletedAt": {
            "$ref": "#/components/schemas/Date"
          }
        },
        "nullable": false,
        "required": [
          "id",
          "title",
          "Tags",
          "content",
          "thumbnail",
          "UserId",
          "User",
          "LikeDisLike",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "BlogPostsTags": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "nullable": false
          },
          "tagName": {
            "type": "string",
            "nullable": false
          },
          "positionType": {
            "type": "string",
            "enum": [
              "Front-End",
              "Back-End",
              "Full-Stack"
            ],
            "nullable": false
          },
          "BlogPosts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlogPosts"
            },
            "nullable": false
          },
          "createdAt": {
            "$ref": "#/components/schemas/Date"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Date"
          },
          "deletedAt": {
            "$ref": "#/components/schemas/Date"
          }
        },
        "nullable": false,
        "required": [
          "id",
          "tagName",
          "positionType",
          "BlogPosts",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "Date": {
        "type": "object",
        "properties": {
          "getVarDate": {}
        },
        "nullable": false,
        "required": [
          "getVarDate"
        ],
        "description": "Enables basic storage and retrieval of dates and times."
      },
      "Users": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "nullable": false
          },
          "githubID": {
            "type": "number",
            "nullable": false
          },
          "email": {
            "type": "string",
            "nullable": false
          },
          "nickname": {
            "type": "string",
            "nullable": false
          },
          "loginID": {
            "type": "string",
            "nullable": false
          },
          "blog": {
            "type": "string",
            "nullable": false
          },
          "bio": {
            "type": "string",
            "nullable": false
          },
          "avatarUrl": {
            "type": "string",
            "nullable": false
          },
          "githubPageUrl": {
            "type": "string",
            "nullable": false
          },
          "password": {
            "type": "string",
            "nullable": false
          },
          "positionType": {
            "type": "string",
            "enum": [
              "Front-End",
              "Back-End",
              "Full-Stack"
            ],
            "nullable": false
          },
          "loginType": {
            "type": "string",
            "enum": [
              "local",
              "github"
            ],
            "nullable": false
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ],
            "nullable": false
          },
          "createdAt": {
            "$ref": "#/components/schemas/Date"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Date"
          },
          "deletedAt": {
            "$ref": "#/components/schemas/Date"
          },
          "Posts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlogPosts"
            },
            "nullable": false
          },
          "LikeDisLike": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlogPostsLike"
            },
            "nullable": false
          }
        },
        "nullable": false,
        "required": [
          "id",
          "githubID",
          "email",
          "nickname",
          "loginID",
          "blog",
          "bio",
          "avatarUrl",
          "githubPageUrl",
          "password",
          "positionType",
          "loginType",
          "role",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "Posts",
          "LikeDisLike"
        ]
      },
      "BlogPostsLike": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "nullable": false
          },
          "actionType": {
            "type": "string",
            "enum": [
              "Like",
              "DisLike"
            ],
            "nullable": false
          },
          "PostId": {
            "type": "number",
            "nullable": false
          },
          "BlogPost": {
            "$ref": "#/components/schemas/BlogPosts"
          },
          "UserId": {
            "type": "number",
            "nullable": false
          },
          "User": {
            "$ref": "#/components/schemas/Users"
          },
          "createdAt": {
            "$ref": "#/components/schemas/Date"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Date"
          },
          "deletedAt": {
            "$ref": "#/components/schemas/Date"
          }
        },
        "nullable": false,
        "required": [
          "id",
          "actionType",
          "PostId",
          "BlogPost",
          "UserId",
          "User",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
...
      "userdata": {
        "type": "object",
        "properties": {},
        "nullable": false
      },
      "LocalSignUpRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": false
          },
          "loginID": {
            "type": "string",
            "nullable": false
          },
          "nickname": {
            "type": "string",
            "nullable": false
          },
          "blog": {
            "type": "string",
            "nullable": false
          },
          "bio": {
            "type": "string",
            "nullable": false
          },
          "password": {
            "type": "string",
            "nullable": false
          },
          "positionType": {
            "type": "string",
            "enum": [
              "Front-End",
              "Back-End",
              "Full-Stack"
            ],
            "nullable": false
          }
        },
        "nullable": false,
        "required": [
          "email",
          "loginID",
          "nickname",
          "blog",
          "bio",
          "password",
          "positionType"
        ]
      },
      "BadRequestException": {
        "type": "object",
        "properties": {},
        "nullable": false,
        "description": "Defines an HTTP exception for *Bad Request* type errors."
      }
    }
  }
}